+++ test/utils/number.test.js
@@ -269,10 +269,8 @@ describe('number', function() {
 
       it('auto notation with custom lower and upper bound', function () {
         var options = {
-          exponential: {
-            lower: 1e-6,
-            upper: 1e+9
-          }
+          lowerExp: -6,
+          upperExp: 9
         };
         assert.equal(number.format(0, options), '0');
         assert.equal(number.format(1234567, options), '1234567');
@@ -282,63 +280,61 @@ describe('number', function() {
         assert.equal(number.format(0.999e-6, options), '9.99e-7');
         assert.equal(number.format(123456789123, options), '1.23456789123e+11');
 
-        assert.equal(number.format(Math.pow(2, 53), {exponential: {upper: 1e+20}}), '9007199254740992');
+        assert.equal(number.format(Math.pow(2, 53), {upperExp: 20}), '9007199254740992');
       });
 
       it('auto notation with custom lower bound', function () {
-        var options = { exponential: { lower: 1e-6 } };
+        var options = { lowerExp: -6 };
         assert.equal(number.format(0, options), '0');
         assert.equal(number.format(1e-6, options), '0.000001');
         assert.equal(number.format(0.999e-6, options), '9.99e-7');
       });
 
       it('auto notation with very large custom lower bound', function () {
-        assert.equal(number.format(1, { exponential: { lower: 1e-2 } }), '1');
-        assert.equal(number.format(1e-1, { exponential: { lower: 1e-2 } }), '0.1');
-        assert.equal(number.format(1e-2, { exponential: { lower: 1e-2} }), '0.01');
-        assert.equal(number.format(1e-3, { exponential: { lower: 1e-2 } }), '1e-3');
+        assert.equal(number.format(1, { lowerExp: -2 }), '1');
+        assert.equal(number.format(1e-1, { lowerExp: -2 }), '0.1');
+        assert.equal(number.format(1e-2, { lowerExp: -2}), '0.01');
+        assert.equal(number.format(1e-3, { lowerExp: -2 }), '1e-3');
       });
 
       it('auto notation with very small custom lower bound', function () {
-        assert.equal(number.format(1e-18, { exponential: { lower: 1e-30 } }), '0.000000000000000001');
-        assert.equal(number.format(1e-19, { exponential: { lower: 1e-30 } }), '0.0000000000000000001');
-        assert.equal(number.format(1e-20, { exponential: { lower: 1e-30 } }), '0.00000000000000000001');
-        assert.equal(number.format(1e-21, { exponential: { lower: 1e-30 } }), '0.000000000000000000001');
-        assert.equal(number.format(1e-22, { exponential: { lower: 1e-30 } }), '0.0000000000000000000001');
-        assert.equal(number.format(1e-23, { exponential: { lower: 1e-30 } }), '0.00000000000000000000001');
-        assert.equal(number.format(1e-24, { exponential: { lower: 1e-30 } }), '0.000000000000000000000001');
+        assert.equal(number.format(1e-18, { lowerExp: -30 }), '0.000000000000000001');
+        assert.equal(number.format(1e-19, { lowerExp: -30 }), '0.0000000000000000001');
+        assert.equal(number.format(1e-20, { lowerExp: -30 }), '0.00000000000000000001');
+        assert.equal(number.format(1e-21, { lowerExp: -30 }), '0.000000000000000000001');
+        assert.equal(number.format(1e-22, { lowerExp: -30 }), '0.0000000000000000000001');
+        assert.equal(number.format(1e-23, { lowerExp: -30 }), '0.00000000000000000000001');
+        assert.equal(number.format(1e-24, { lowerExp: -30 }), '0.000000000000000000000001');
       });
 
       it('auto notation with custom upper bound', function () {
-        var options = { exponential: { upper: 1e+9 } };
+        var options = { upperExp: 9 };
         assert.equal(number.format(1e+9, options), '1e+9');
         assert.equal(number.format(1e+9-1, options), '999999999');
       });
 
       it('auto notation with very large custom upper bound', function () {
-        assert.equal(number.format(1e+18, { exponential: { upper: 1e+30 } }), '1000000000000000000');
-        assert.equal(number.format(1e+19, { exponential: { upper: 1e+30 } }), '10000000000000000000');
-        assert.equal(number.format(1e+20, { exponential: { upper: 1e+30 } }), '100000000000000000000');
-        assert.equal(number.format(1e+21, { exponential: { upper: 1e+30 } }), '1000000000000000000000');
-        assert.equal(number.format(1e+22, { exponential: { upper: 1e+30 } }), '10000000000000000000000');
-        assert.equal(number.format(1e+23, { exponential: { upper: 1e+30 } }), '100000000000000000000000');
-        assert.equal(number.format(1e+24, { exponential: { upper: 1e+30 } }), '1000000000000000000000000');
+        assert.equal(number.format(1e+18, { upperExp: 30 }), '1000000000000000000');
+        assert.equal(number.format(1e+19, { upperExp: 30 }), '10000000000000000000');
+        assert.equal(number.format(1e+20, { upperExp: 30 }), '100000000000000000000');
+        assert.equal(number.format(1e+21, { upperExp: 30 }), '1000000000000000000000');
+        assert.equal(number.format(1e+22, { upperExp: 30 }), '10000000000000000000000');
+        assert.equal(number.format(1e+23, { upperExp: 30 }), '100000000000000000000000');
+        assert.equal(number.format(1e+24, { upperExp: 30 }), '1000000000000000000000000');
       });
 
       it('auto notation with very small custom upper bound', function () {
-        assert.equal(number.format(1, { exponential: { upper: 1e2 } }), '1');
-        assert.equal(number.format(1e1, { exponential: { upper: 1e2 } }), '10');
-        assert.equal(number.format(1e2, { exponential: { upper: 1e2 } }), '1e+2');
-        assert.equal(number.format(1e3, { exponential: { upper: 1e2 } }), '1e+3');
+        assert.equal(number.format(1, { upperExp: 2 }), '1');
+        assert.equal(number.format(1e1, { upperExp: 2 }), '10');
+        assert.equal(number.format(1e2, { upperExp: 2 }), '1e+2');
+        assert.equal(number.format(1e3, { upperExp: 2 }), '1e+3');
       });
 
       it('auto notation with custom precision, lower, and upper bound', function () {
         var options = {
           precision: 4,
-          exponential: {
-            lower: 1e-6,
-            upper: 1e+9
-          }
+          lowerExp: -6,
+          upperExp: 9
         };
 
         assert.equal(number.format(0, options), '0');